Add public profile settings page and refactor profile management.
- Introduced `MyProfileSettingsPage` for managing display name and avatar shown on public leaderboards.
- Refactored `MyProfilePage` to remove public profile management functionality and link to settings page.
- Updated lazy loading and route configuration for `MyProfileSettingsPage` with feature flag `user_login`.
- Enhanced i18n files with strings for public profile settings.
